(ui-update):options card

+++ frontend/src/adminComponents/StaffCard.jsx
@@ -22,6 +22,7 @@ const StaffCard = ({ user, refetch }) => {
 
   const updateBlock = async () => {
     try {
+      if (deleteLoading || updateLoading) return;
       setBlockLoading(true);
       let response = await fetch(`${BACKEND_URL}/api/users/${user._id}`, {
         method: "PATCH",
@@ -45,7 +46,7 @@ const StaffCard = ({ user, refetch }) => {
 
   const deleteUser = async () => {
     try {
-      if (deleteLoading) return;
+      if (deleteLoading || updateLoading) return;
       setDeleteLoading(true);
       let response = await fetch(`${BACKEND_URL}/api/users/${user._id}`, {
         method: "DELETE",
@@ -109,9 +110,9 @@ const StaffCard = ({ user, refetch }) => {
               )}
             </button>
             <button
-              className={` py-2 px-4 text-red-700 hover:bg-neutral-100 transition-colors ${deleteLoading && "cursor-not-allowed opacity-70"}`}
+              className={` py-2 px-4 text-red-700 hover:bg-neutral-100 transition-colors ${(deleteLoading || updateLoading) && "cursor-not-allowed opacity-70"}`}
               onClick={deleteUser}
-              disabled={deleteLoading}
+              disabled={deleteLoading || updateLoading}
             >
               {deleteLoading ? (
                 <div className="flex items-center gap-1">
